removed old booker and make new booker as a default
