+++ .env.example
@@ -214,9 +214,3 @@ PROJECT_ID_VERCEL=
 TEAM_ID_VERCEL=
 # Get it from: https://vercel.com/account/tokens
 AUTH_BEARER_TOKEN_VERCEL=
-
-#Enables New booker for Embed only
-NEW_BOOKER_ENABLED_FOR_EMBED=0
-
-#Enables New booker for All but Embed requests
-NEW_BOOKER_ENABLED_FOR_NON_EMBED=0
