+++ apps/web/pages/d/[link]/[slug].tsx
@@ -1,42 +1,50 @@
 import type { GetServerSidePropsContext } from "next";
 import { z } from "zod";
 
-import type { LocationObject } from "@calcom/core/location";
-import { privacyFilteredLocations } from "@calcom/core/location";
+import { Booker } from "@calcom/atoms";
+import { BookerSeo } from "@calcom/features/bookings/components/BookerSeo";
+import { getBookingByUidOrRescheduleUid } from "@calcom/features/bookings/lib/get-booking";
 import type { GetBookingType } from "@calcom/features/bookings/lib/get-booking";
-import { parseRecurringEvent } from "@calcom/lib";
-import { getWorkingHours } from "@calcom/lib/availability";
-import { markdownToSafeHTML } from "@calcom/lib/markdownToSafeHTML";
-import { availiblityPageEventTypeSelect } from "@calcom/prisma";
+import { orgDomainConfig } from "@calcom/features/ee/organizations/lib/orgDomains";
+import slugify from "@calcom/lib/slugify";
 import prisma from "@calcom/prisma";
-import { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
 
 import type { inferSSRProps } from "@lib/types/inferSSRProps";
-import type { EmbedProps } from "@lib/withEmbedSsr";
 
 import PageWrapper from "@components/PageWrapper";
-import AvailabilityPage from "@components/booking/pages/AvailabilityPage";
 
-import { ssrInit } from "@server/lib/ssr";
+type PageProps = inferSSRProps<typeof getServerSideProps>;
 
-export type DynamicAvailabilityPageProps = inferSSRProps<typeof getServerSideProps> & EmbedProps;
-
-export default function Type(props: DynamicAvailabilityPageProps) {
-  return <AvailabilityPage {...props} />;
+export default function Type({ slug, user, booking, away, isBrandingHidden, isTeamEvent }: PageProps) {
+  return (
+    <main className="flex h-full min-h-[100dvh] items-center justify-center">
+      <BookerSeo
+        username={user}
+        eventSlug={slug}
+        rescheduleUid={booking?.uid}
+        hideBranding={isBrandingHidden}
+      />
+      <Booker
+        username={user}
+        eventSlug={slug}
+        rescheduleBooking={booking}
+        isAway={away}
+        hideBranding={isBrandingHidden}
+        isTeamEvent={isTeamEvent}
+      />
+    </main>
+  );
 }
 
-Type.isBookingPage = true;
 Type.PageWrapper = PageWrapper;
 
-const querySchema = z.object({
-  link: z.string().optional().default(""),
-  slug: z.string().optional().default(""),
-  date: z.union([z.string(), z.null()]).optional().default(null),
-});
+async function getUserPageProps(context: GetServerSidePropsContext) {
+  const { link, slug } = paramsSchema.parse(context.params);
+  const { rescheduleUid } = context.query;
+  const { currentOrgDomain, isValidOrgDomain } = orgDomainConfig(context.req.headers.host ?? "");
 
-export const getServerSideProps = async (context: GetServerSidePropsContext) => {
+  const { ssrInit } = await import("@server/lib/ssr");
   const ssr = await ssrInit(context);
-  const { link, slug, date } = querySchema.parse(context.query);
 
   const hashedLink = await prisma.hashedLink.findUnique({
     where: {
@@ -45,132 +53,87 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
     select: {
       eventTypeId: true,
       eventType: {
-        select: availiblityPageEventTypeSelect,
-      },
-    },
-  });
-
-  const userId = hashedLink?.eventType.userId || hashedLink?.eventType.users[0]?.id;
-  if (!userId)
-    return {
-      notFound: true,
-    } as {
-      notFound: true;
-    };
-
-  if (hashedLink?.eventType.slug !== slug)
-    return {
-      notFound: true,
-    } as {
-      notFound: true;
-    };
-
-  const users = await prisma.user.findMany({
-    where: {
-      id: userId,
-    },
-    select: {
-      id: true,
-      username: true,
-      name: true,
-      email: true,
-      bio: true,
-      avatar: true,
-      startTime: true,
-      endTime: true,
-      timeZone: true,
-      weekStart: true,
-      availability: true,
-      hideBranding: true,
-      brandColor: true,
-      darkBrandColor: true,
-      defaultScheduleId: true,
-      allowDynamicBooking: true,
-      away: true,
-      schedules: {
         select: {
-          availability: true,
-          timeZone: true,
-          id: true,
+          users: {
+            select: {
+              username: true,
+            },
+          },
+          team: {
+            select: {
+              id: true,
+            },
+          },
         },
       },
-      theme: true,
     },
   });
 
-  if (!users || !users.length) {
+  const username = hashedLink?.eventType.users[0]?.username;
+
+  if (!hashedLink || !username) {
     return {
       notFound: true,
-    } as {
-      notFound: true;
     };
   }
 
-  const locations = hashedLink.eventType.locations
-    ? (hashedLink.eventType.locations as LocationObject[])
-    : [];
-
-  const eventTypeObject = Object.assign({}, hashedLink.eventType, {
-    metadata: EventTypeMetaDataSchema.parse(hashedLink.eventType.metadata || {}),
-    recurringEvent: parseRecurringEvent(hashedLink.eventType.recurringEvent),
-    periodStartDate: hashedLink.eventType.periodStartDate?.toString() ?? null,
-    periodEndDate: hashedLink.eventType.periodEndDate?.toString() ?? null,
-    slug,
-    locations: privacyFilteredLocations(locations),
-    users: users.map((u) => ({
-      name: u.name,
-      username: u.username,
-      hideBranding: u.hideBranding,
-      timeZone: u.timeZone,
-    })),
-    descriptionAsSafeHTML: markdownToSafeHTML(hashedLink.eventType.description),
+  const user = await prisma.user.findFirst({
+    where: {
+      username,
+      organization: isValidOrgDomain
+        ? {
+            slug: currentOrgDomain,
+          }
+        : null,
+    },
+    select: {
+      away: true,
+      hideBranding: true,
+    },
   });
 
-  const [user] = users;
+  if (!user) {
+    return {
+      notFound: true,
+    };
+  }
 
-  const schedule = {
-    ...user.schedules.filter(
-      (schedule) => !user.defaultScheduleId || schedule.id === user.defaultScheduleId
-    )[0],
-  };
+  let booking: GetBookingType | null = null;
+  if (rescheduleUid) {
+    booking = await getBookingByUidOrRescheduleUid(`${rescheduleUid}`);
+  }
 
-  const timeZone = schedule.timeZone || user.timeZone;
+  const isTeamEvent = !!hashedLink.eventType?.team?.id;
 
-  const workingHours = getWorkingHours(
-    {
-      timeZone,
-    },
-    schedule.availability || user.availability
-  );
-  eventTypeObject.schedule = null;
-  eventTypeObject.availability = [];
+  // We use this to both prefetch the query on the server,
+  // as well as to check if the event exist, so we c an show a 404 otherwise.
+  const eventData = await ssr.viewer.public.event.fetch({ username, eventSlug: slug, isTeamEvent });
 
-  const booking: GetBookingType | null = null;
-
-  const profile = {
-    name: user.name || user.username,
-    image: user.avatar,
-    slug: user.username,
-    theme: user.theme,
-    weekStart: user.weekStart,
-    brandColor: user.brandColor,
-    darkBrandColor: user.darkBrandColor,
-  };
+  if (!eventData) {
+    return {
+      notFound: true,
+    };
+  }
 
   return {
     props: {
-      away: user.away,
-      themeBasis: user.username,
-      isDynamicGroup: false,
-      profile,
-      date,
-      eventType: eventTypeObject,
-      workingHours,
-      trpcState: ssr.dehydrate(),
-      previousPage: context.req.headers.referer ?? null,
       booking,
-      users: [user.username],
-      isBrandingHidden: user.hideBranding,
+      away: user?.away,
+      user: username,
+      slug,
+      trpcState: ssr.dehydrate(),
+      isBrandingHidden: user?.hideBranding,
+      // Sending the team event from the server, because this template file
+      // is reused for both team and user events.
+      isTeamEvent,
     },
   };
+}
+
+const paramsSchema = z.object({ link: z.string(), slug: z.string().transform((s) => slugify(s)) });
+
+// Booker page fetches a tiny bit of data server side, to determine early
+// whether the page should show an away state or dynamic booking not allowed.
+export const getServerSideProps = async (context: GetServerSidePropsContext) => {
+  return await getUserPageProps(context);
 };
