--- apps/web/pages/new-booker/[user]/embed.tsx
@@ -1,7 +1,0 @@
-import withEmbedSsr from "@lib/withEmbedSsr";
-
-import { getServerSideProps as _getServerSideProps } from "../../[user]";
-
-export { default } from "../../[user]";
-
-export const getServerSideProps = withEmbedSsr(_getServerSideProps);
