--- apps/web/pages/new-booker/team/[slug]/[type]/embed.tsx
@@ -1,9 +1,0 @@
-import withEmbedSsr from "@lib/withEmbedSsr";
-
-import { getServerSideProps as _getServerSideProps } from "../[type]";
-
-export { default } from "../[type]";
-
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-export const getServerSideProps = withEmbedSsr(_getServerSideProps);
