+++ apps/web/pages/team/[slug]/[type]/embed.tsx
@@ -4,4 +4,6 @@ import { getServerSideProps as _getServerSideProps } from "../[type]";
 
 export { default } from "../[type]";
 
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 export const getServerSideProps = withEmbedSsr(_getServerSideProps);
