+++ apps/web/pages/team/[slug]/embed.tsx
@@ -1,7 +1,7 @@
 import withEmbedSsr from "@lib/withEmbedSsr";
 
-import { getServerSideProps as _getServerSideProps } from "../[slug]";
+import { getServerSideProps as _getServerSideProps } from "../../../team/[slug]";
 
-export { default } from "../[slug]";
+export { default } from "../../../team/[slug]";
 
 export const getServerSideProps = withEmbedSsr(_getServerSideProps);
