+++ apps/web/playwright/booking-pages.e2e.ts
@@ -1,7 +1,6 @@
 import { expect } from "@playwright/test";
 
 import { test } from "./lib/fixtures";
-import { testBothBookers } from "./lib/new-booker";
 import {
   bookFirstEvent,
   bookOptinEvent,
@@ -15,7 +14,7 @@ import {
 test.describe.configure({ mode: "parallel" });
 test.afterEach(async ({ users }) => users.deleteAll());
 
-testBothBookers.describe("free user", (bookerVariant) => {
+test.describe("free user", () => {
   test.beforeEach(async ({ page, users }) => {
     const free = await users.create();
     await page.goto(`/${free.username}`);
@@ -26,17 +25,6 @@ testBothBookers.describe("free user", (bookerVariant) => {
     await page.click('[data-testid="event-type-link"]');
 
     await selectFirstAvailableTimeSlotNextMonth(page);
-
-    // Kept in if statement here, since it's only temporary
-    // until the old booker isn't used anymore, and I wanted
-    // to change the test as little as possible.
-    // eslint-disable-next-line playwright/no-conditional-in-test
-    if (bookerVariant !== "new-booker") {
-      // Navigate to book page
-      await page.waitForURL((url) => {
-        return url.pathname.endsWith("/book");
-      });
-    }
 
     // save booking url
     const bookingUrl: string = page.url();
@@ -58,7 +46,7 @@ testBothBookers.describe("free user", (bookerVariant) => {
   });
 });
 
-testBothBookers.describe("pro user", () => {
+test.describe("pro user", () => {
   test.beforeEach(async ({ page, users }) => {
     const pro = await users.create();
     await page.goto(`/${pro.username}`);
