+++ apps/web/playwright/event-types.e2e.ts
@@ -4,13 +4,12 @@ import { WEBAPP_URL } from "@calcom/lib/constants";
 import { randomString } from "@calcom/lib/random";
 
 import { test } from "./lib/fixtures";
-import { testBothBookers } from "./lib/new-booker";
 import { bookTimeSlot, createNewEventType, selectFirstAvailableTimeSlotNextMonth } from "./lib/testUtils";
 
 test.describe.configure({ mode: "parallel" });
 
 test.describe("Event Types tests", () => {
-  testBothBookers.describe("user", (bookerVariant) => {
+  test.describe("user", () => {
     test.beforeEach(async ({ page, users }) => {
       const user = await users.create();
       await user.apiLogin();
@@ -142,17 +141,6 @@ test.describe("Event Types tests", () => {
       await page.goto(previewLink ?? "");
 
       await selectFirstAvailableTimeSlotNextMonth(page);
-
-      // Navigate to book page
-      // Kept in if statement here, since it's only temporary
-      // until the old booker isn't used anymore, and I wanted
-      // to change the test as little as possible.
-      // eslint-disable-next-line playwright/no-conditional-in-test
-      if (bookerVariant === "old-booker") {
-        await page.waitForURL((url) => {
-          return url.pathname.endsWith("/book");
-        });
-      }
 
       for (const location of locationData) {
         await page.locator(`span:has-text("${location}")`).click();
