+++ apps/web/playwright/manage-booking-questions.e2e.ts
@@ -7,8 +7,6 @@ import prisma from "@calcom/prisma";
 import { WebhookTriggerEvents } from "@calcom/prisma/enums";
 
 import { test } from "./lib/fixtures";
-import { testBothBookers } from "./lib/new-booker";
-import type { BookerVariants } from "./lib/new-booker";
 import { createHttpServer, waitFor, selectFirstAvailableTimeSlotNextMonth } from "./lib/testUtils";
 
 async function getLabelText(field: Locator) {
@@ -21,7 +19,7 @@ test.describe("Manage Booking Questions", () => {
     await users.deleteAll();
   });
 
-  testBothBookers.describe("For User EventType", (bookerVariant) => {
+  test.describe("For User EventType", () => {
     test("Do a booking with a user added question and verify a few thing in b/w", async ({
       page,
       users,
@@ -40,11 +38,11 @@ test.describe("Manage Booking Questions", () => {
         await firstEventTypeElement.click();
       });
 
-      await runTestStepsCommonForTeamAndUserEventType(page, context, webhookReceiver, bookerVariant);
+      await runTestStepsCommonForTeamAndUserEventType(page, context, webhookReceiver);
     });
   });
 
-  testBothBookers.describe("For Team EventType", (bookerVariant) => {
+  test.describe("For Team EventType", () => {
     test("Do a booking with a user added question and verify a few thing in b/w", async ({
       page,
       users,
@@ -76,7 +74,7 @@ test.describe("Manage Booking Questions", () => {
         await firstEventTypeElement.click();
       });
 
-      await runTestStepsCommonForTeamAndUserEventType(page, context, webhookReceiver, bookerVariant);
+      await runTestStepsCommonForTeamAndUserEventType(page, context, webhookReceiver);
     });
   });
 });
@@ -90,8 +88,7 @@ async function runTestStepsCommonForTeamAndUserEventType(
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     requestList: (import("http").IncomingMessage & { body?: any })[];
     url: string;
-  },
-  bookerVariant: BookerVariants
+  }
 ) {
   await page.click('[href$="tabName=advanced"]');
   await test.step("Add Question and see that it's shown on Booking Page at appropriate position", async () => {
@@ -106,7 +103,7 @@ async function runTestStepsCommonForTeamAndUserEventType(
       },
     });
 
-    await doOnFreshPreview(page, context, bookerVariant, async (page) => {
+    await doOnFreshPreview(page, context, async (page) => {
       const allFieldsLocator = await expectSystemFieldsToBeThere(page);
       const userFieldLocator = allFieldsLocator.nth(5);
 
@@ -122,7 +119,7 @@ async function runTestStepsCommonForTeamAndUserEventType(
       name: "how_are_you",
       page,
     });
-    await doOnFreshPreview(page, context, bookerVariant, async (page) => {
+    await doOnFreshPreview(page, context, async (page) => {
       const formBuilderFieldLocator = page.locator('[data-fob-field-name="how_are_you"]');
       await expect(formBuilderFieldLocator).toBeHidden();
     });
@@ -136,7 +133,7 @@ async function runTestStepsCommonForTeamAndUserEventType(
   });
 
   await test.step('Try to book without providing "How are you?" response', async () => {
-    await doOnFreshPreview(page, context, bookerVariant, async (page) => {
+    await doOnFreshPreview(page, context, async (page) => {
       await bookTimeSlot({ page, name: "Booker", email: "booker@example.com" });
       await expectErrorToBeThereFor({ page, name: "how_are_you" });
     });
@@ -155,7 +152,6 @@ async function runTestStepsCommonForTeamAndUserEventType(
       return await doOnFreshPreview(
         page,
         context,
-        bookerVariant,
         async (page) => {
           const formBuilderFieldLocator = page.locator('[data-fob-field-name="how_are_you"]');
           await expect(formBuilderFieldLocator).toBeVisible();
@@ -324,11 +320,10 @@ async function expectErrorToBeThereFor({ page, name }: { page: Page; name: strin
 async function doOnFreshPreview(
   page: Page,
   context: PlaywrightTestArgs["context"],
-  bookerVariant: BookerVariants,
   callback: (page: Page) => Promise<void>,
   persistTab = false
 ) {
-  const previewTabPage = await openBookingFormInPreviewTab(context, page, bookerVariant);
+  const previewTabPage = await openBookingFormInPreviewTab(context, page);
   await callback(previewTabPage);
   if (!persistTab) {
     await previewTabPage.close();
@@ -377,39 +372,19 @@ async function createAndLoginUserWithEventTypes({
   return user;
 }
 
-async function rescheduleFromTheLinkOnPage({
-  page,
-  bookerVariant,
-}: {
-  page: Page;
-  bookerVariant?: BookerVariants;
-}) {
+async function rescheduleFromTheLinkOnPage({ page }: { page: Page }) {
   await page.locator('[data-testid="reschedule-link"]').click();
   await page.waitForLoadState();
   await selectFirstAvailableTimeSlotNextMonth(page);
-  if (bookerVariant === "old-booker") {
-    await page.waitForURL((url) => {
-      return url.pathname.endsWith("/book");
-    });
-  }
   await page.click('[data-testid="confirm-reschedule-button"]');
 }
 
-async function openBookingFormInPreviewTab(
-  context: PlaywrightTestArgs["context"],
-  page: Page,
-  bookerVariant: BookerVariants
-) {
+async function openBookingFormInPreviewTab(context: PlaywrightTestArgs["context"], page: Page) {
   const previewTabPromise = context.waitForEvent("page");
   await page.locator('[data-testid="preview-button"]').click();
   const previewTabPage = await previewTabPromise;
   await previewTabPage.waitForLoadState();
   await selectFirstAvailableTimeSlotNextMonth(previewTabPage);
-  if (bookerVariant === "old-booker") {
-    await previewTabPage.waitForURL((url) => {
-      return url.pathname.endsWith("/book");
-    });
-  }
   return previewTabPage;
 }
