+++ turbo.json
@@ -214,8 +214,6 @@
     "LARK_OPEN_VERIFICATION_TOKEN",
     "MS_GRAPH_CLIENT_ID",
     "MS_GRAPH_CLIENT_SECRET",
-    "NEW_BOOKER_ENABLED_FOR_EMBED",
-    "NEW_BOOKER_ENABLED_FOR_NON_EMBED",
     "NEXT_PUBLIC_API_URL",
     "NEXT_PUBLIC_APP_NAME",
     "NEXT_PUBLIC_AUTH_URL",
